+++ app/src/main/java/org/stepic/droid/di/course_list/CourseListComponent.kt
@@ -4,7 +4,7 @@ import dagger.Subcomponent
 import org.stepic.droid.base.CoursesDatabaseFragmentBase
 import org.stepic.droid.ui.fragments.CourseListFragmentBase
 import org.stepic.droid.ui.fragments.CourseSearchFragment
-import org.stepic.droid.ui.fragments.HomeFragment
+import org.stepic.droid.ui.fragments.CoursesCarouselFragment
 
 @CourseListScope
 @Subcomponent(modules = arrayOf(CourseListModule::class))
@@ -22,5 +22,5 @@ interface CourseListComponent {
 
     fun inject(fragment: CourseSearchFragment)
 
-    fun inject(fragment: HomeFragment)
+    fun inject(fragment: CoursesCarouselFragment)
 }
