+++ app/src/main/res/layout/fragment_courses_carousel.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.v7.widget.RecyclerView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/myCoursesOnHome"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:clipToPadding="false"
+    android:divider="@color/transparent"
+    android:overScrollMode="never"
+    android:paddingRight="@dimen/home_right_recycler_padding"
+    android:scrollbarStyle="outsideOverlay"
+    android:scrollbars="none"
+    android:splitMotionEvents="false"/>
+
