+++ app/src/main/res/layout/fragment_home.xml
@@ -1,31 +1,39 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.design.widget.CoordinatorLayout
+    android:id="@+id/homeRootView"
+    android:background="@color/white"
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
     <include
         android:id="@+id/appBarLayout"
         layout="@layout/view_centered_appbar"/>
 
-    <FrameLayout
+    <android.support.v4.widget.NestedScrollView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:background="@color/white"
         app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
-        <android.support.v7.widget.RecyclerView
-            android:id="@+id/myCoursesOnHome"
+        <LinearLayout
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:clipToPadding="false"
-            android:divider="@color/transparent"
-            android:overScrollMode="never"
-            android:paddingRight="@dimen/home_right_recycler_padding"
-            android:scrollbarStyle="outsideOverlay"
-            android:scrollbars="none"
-            android:splitMotionEvents="false"/>
-    </FrameLayout>
+            android:layout_height="match_parent"
+            android:orientation="vertical">
 
+            <fragment
+                android:id="@+id/myCourses"
+                android:name="org.stepic.droid.ui.fragments.CoursesCarouselFragment"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"/>
+
+            <fragment
+                android:id="@+id/popularCourses"
+                android:name="org.stepic.droid.ui.fragments.CoursesCarouselFragment"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"/>
+
+        </LinearLayout>
+    </android.support.v4.widget.NestedScrollView>
 </android.support.design.widget.CoordinatorLayout>
